Run 'git log --shortstat' twice iff necessary
Because of the way the code is written now, 'git log --shortstat'
is executed twice. Once to fetch line statistics, the other time
for fetching per=author statistics. Well, since the output of git
log is not changed in between the runs, there's no need to run
this expensive command twice.

The only exception is when linear_linestats are enabled. Then the
line stats cmd line is different and we have to run it.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>